[FIX] Binary and libhl.so matches official Makefile target name.
Previously we use target "hl" to build libhl.so, and "hashlink" to build
executable. This is inconsistent with official release. At that time
I pick this naming because I was restricted by xmake, that it does not
support two targets defined with same name.

After a study I found it's indeed possible by calling
target:set_basename() function in target of libhl.so. Thus I can use
consistent target name with Makefile.