[FIX] Finally make openal-soft becomes static!
NOTE: In theory openal-soft should not be static due to its nature of
LGPL-2.0 licensing issue. However, it's a C++ library, and it causes
libstdc++.so versioning problem, which can break execution when moving
to different Linux distros.

We build it by-default as static linked library. However it may cause
problems on licensing. Use with caution.