[NEW] Finally fixed crash issue.
Root cause is LIBHL_EXPORTS macro should be used only when compling
libhl.dll, not all modules. If we apply the macro on all modules, a
name resolving conflict happens with a misleading message, that it can
resolve symbol. The true problem is the symbols (hlt_*) is exported
multiple times, leading to a crash when calling some functions.