[NEW] Linux: Copy alsa-lib and libsndio as dependencies.
We have to build OpenAL-soft as shared library because the
structure of Hashlink uses alGetProcAddress() API (similar with
dlopen()) to load extension APIs. Thus, linking OpenAL as static library
causes dup-symbol link error.

Since we have to build OpenAL in shared library, we need to ensure its
dependencies, alsa-lib and libsndio, are built as shared libraries too,
so they can be copied to release package (xmake install) as well. To do
this, all three libraries are configured to build from source, as shared
libraries. However, in release package we have to separate libopenal.so,
in order to use the system-provided OpenAL.

This can still break in some systems, but it's the best idea I can find
so far to build a portable binary set.