[NEW] Build openal.hdll statically (#5)
Openal.hdll was not able to build statically because of the existing of libsndio. By default libsndio requires building as a shared library, thus, it breaks the dependency chain (hashlink::openal.hdll -> openal-soft -> libsndio -> alsa-lib) when trying to build with -static-libgcc and -static-libstdc++.

There are two possible fixes: a) convert libsndio as a static library, and b) add -static-libgcc and -static-libstdc++ to libsndio.so. The final solution I chose is a). It requires three steps:

    Provide a private package for libsndio to allow it build a static libsndio-static.a.
    Note that this is not accepted by official.
    Change my own xmake.lua. Use set_requireconfs() to require the full chain building in static library.
    Apply openal.hdll patch to avoid static library link error. Note this is not accepted by official Hashlink.

With the changes, we ends up with a libopenal.a, which depends on only libhl.so, libm.so.6, libc.so.6, and ld-linux-x86-64.so.2. The risky libraries, libgcc_s.so.6 and libstdc++.so.6, are embeded to openal.hdll.

The reason we don't use solution b) is, when libsndio is a shared library, then openal-soft needs to be a shared library as well. openal-soft is a C++ library, which explicitly requires libstdc++.so.6. This is too risky for portability.


Why is a private path required for openal.c:

Hashlink applies a trick to load openal-soft functions. The idea is to
define a set of function pointers via AL_IMPORT and ALC_IMPORT macro,
then use a ``func_name = alGetProcAddress("func_name")`` to initialize
function pointers, which appears no difference between a dynamic loading
and directly linking. This trick is applied when loading extension
functions.

This trick, however, causes a duplicated symbol link error when
linking against openal-al static library, because the function pointers
(from openal.hdll) and real functions (from libopenal.a) has naming
conflict.

The fix is to change the generated function pointer naming pattern,
adding a _dyn suffix instead of directly reusing function names.